requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
bleach==3.2.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
ffmpeg-python==0.2.0
fire==0.3.1
future==0.18.2
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-speech==2.0.1
google-cloud-storage==1.35.0
google-cloud-texttospeech==2.2.0
google-cloud-translate==3.0.2
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
iniconfig==1.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
isort==5.7.0
jedi==0.17.2
jenkspy==0.2.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
libcst==0.3.16
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
moviepy==1.0.3
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.6
numpy==1.19.5
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
proglog==0.1.9
prometheus-client==0.9.0
prompt-toolkit==3.0.9
proto-plus==1.13.0
protobuf==3.14.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydub==0.24.1
Pygments==2.7.3
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.1
rsa==4.6
send2trash==1.5.0
six==1.15.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
asgiref==3.2.10
beautifulsoup4==4.9.2
Django==3.0.7
django-bootstrap4==1.1.1
gunicorn==20.0.4
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.2.0
boto3==1.18.48
botocore==1.21.48; python_version >= '3.6'
django-storages==1.11.1